refact: move render stuff to seprate file
